Move stuff to graph_api subdirectory
