+++ caldir-provider-outlook/src/graph_api/client.rs
