+++ caldir-provider-outlook/src/graph_api/mod.rs
@@ -1,0 +1,2 @@
+pub mod client;
+pub mod types;
