+++ caldir-provider-outlook/src/graph_api/types.rs
